fix : 진료 및 처방 관련 레이아웃 동작 수정
 - 진료 선택 싱크 관련 로직 개선
 - 진료 처방내역 레이아웃 수정
 - 진료 오더목록 상단 섹션에서 오더 선택시 동작 구현
 - 선택여부에 따른 제어를 위한 ToggleButton 구현

+++ SmartEMR.Application/Views/SmartEMRCST/vSmartEMRConsultationTabCSTOInfo.xaml.cs
@@ -45,7 +45,11 @@ public partial class vSmartEMRConsultationTabCSTOInfo : ModelViewLayout<Consulta
         switch (e.Column.FieldName)
         {
             case "btnDelete":
-                DeleteCSTO(dataItem);
+                if (SmartUI.MsgYesNo("삭제하시겠습니까?") is System.Windows.MessageBoxResult.Yes)
+                {
+                    DeleteCSTO(dataItem);
+                }
+
                 break;
         }
     }
@@ -65,7 +69,7 @@ public partial class vSmartEMRConsultationTabCSTOInfo : ModelViewLayout<Consulta
         vm.AddCSTO(item);
     }
 
-    private void DeleteCSTO(ConsultationOrder item)
+    public void DeleteCSTO(ConsultationOrder item)
     {
         vm.DeleteCSTO(item);
     }
